refactor(labware-library): Update TableTitle expandable styles
