+++ labware-library/src/components/ui/LabeledValueTable.js
@@ -1,13 +1,10 @@
 // @flow
 import * as React from 'react'
-import cx from 'classnames'
 
 import { Table, TableEntry, TABLE_COLUMN } from './Table'
 import { LabelText, LABEL_LEFT } from './LabelText'
 import { Value } from './Value'
-import { ClickableIcon } from './ClickableIcon'
-
-import styles from './styles.css'
+import { TableTitle } from './TableTitle'
 
 import type { TableDirection } from './Table'
 
@@ -39,40 +36,6 @@ export function LabeledValueTable(props: LabledValueTableProps) {
         ))}
       </Table>
       {children}
-    </div>
-  )
-}
-
-type TableTitleProps = {|
-  label: React.Node,
-  diagram?: React.Node,
-|}
-
-export function TableTitle(props: TableTitleProps) {
-  const [guideVisible, setGuideVisible] = React.useState<boolean>(false)
-  const toggleGuide = () => setGuideVisible(!guideVisible)
-  const { label, diagram } = props
-
-  const iconClassName = cx(styles.info_button, {
-    [styles.active]: guideVisible,
-  })
-
-  const contentClassName = cx(styles.expandable_content, {
-    [styles.open]: guideVisible,
-  })
-
-  return (
-    <div className={styles.table_title}>
-      <div className={styles.table_title_text}>
-        <LabelText position={LABEL_LEFT}>{label}</LabelText>
-        <ClickableIcon
-          title="info"
-          name="information"
-          className={iconClassName}
-          onClick={toggleGuide}
-        />
-      </div>
-      <div className={contentClassName}>{diagram}</div>
     </div>
   )
 }
